Fix up the tap/tun header file detection on freebsd-10.
That header uses u_char as a type but it's not defined by default.
Thus pull in sys/types.h if it's there; that will find it on FreeBSD-10.

Add the library to the build.

This still doesn't do everything; it turns out config.h is already a file
so I need to rename this one to cmake_config.h and make including it
optional.