Add a method to clear a signal handler.
SETSIG() does a longjump hack to reset the call flow back to the given
function to execute (rather than a callback? Sigh :-)

However the shutdown call that resets the filedescriptor actually
signals kEABORT, which unfortunately does silly things in the
current processing path.  (And I haven't tried 'reset' on a normal
telnet/connect session; I bet it's going to behave poorly.)
kEABORT was only supposed to be used for connect failures, not session
close/abort once it was connected.

So, add a function to clear the given signal handler so consumers
can eventually just call CLEARSIG() to avoid being signaled for things
it isn't ready for.